Scale ButtonBar fonts and icons based on zoom level, not window size [v3]
This makes the whole UI get bigger or smaller depending on the zoom
level. Icons are scaled by the screen resolution as well so that they
will always appear in the same size relative to all fonts.

This provides user control over appearance, which should make the
interface more accessible for users with low vision and across a
wide range of displays.

Signed-off-by: Keith Packard <keithp@keithp.com>

---

v2:
	Fix tests

		Remove test_Window_resizeEvent as the button bar no
		longer responds to window size changes

		Replace test_ButtonBar_set_responsive_mode with
		test_ButtonBar_set_zoom. This replacement test
		verifies that changing the zoom level adjusts the font
		and icon sizes.

	Remove DEFAULT_FONT_SIZE import from mu/interface/main.py

		This is no longer used

v3
	Use mock for QDesktopWidget in zoom tests

Signed-off-by: Keith Packard <keithp@keithp.com>